feat(core): move refresh code from utilities to its own file
